backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
pyswip>=0.3.0,<1.0
psycopg2-binary>=2.9,<3.0
